+++ evm-template/node/src/service.rs
@@ -21,8 +21,9 @@ use cumulus_relay_chain_interface::{OverseerHandle, RelayChainInterface};
 use frame_benchmarking_cli::SUBSTRATE_REFERENCE_HARDWARE;
 // Local Runtime Types
 use parachain_template_runtime::{
+    apis::RuntimeApi,
     opaque::{Block, Hash},
-    RuntimeApi, TransactionConverter,
+    TransactionConverter,
 };
 use sc_client_api::Backend;
 use sc_consensus::ImportQueue;
@@ -48,7 +49,7 @@ impl sc_executor::NativeExecutionDispatch for ParachainNativeExecutor {
     type ExtendHostFunctions = frame_benchmarking::benchmarking::HostFunctions;
 
     fn dispatch(method: &str, data: &[u8]) -> Option<Vec<u8>> {
-        parachain_template_runtime::api::dispatch(method, data)
+        parachain_template_runtime::apis::api::dispatch(method, data)
     }
 
     fn native_version() -> sc_executor::NativeVersion {
